# Disable reference counting for constants