--- a/compiler/cli/src/run.rs
+++ b/compiler/cli/src/run.rs
@@ -72,9 +72,9 @@ pub(crate) fn run(options: Options) -> ProgramResult {
         ("Stdout", SendPort::create(&mut ended.heap, stdout.channel)),
         ("Stdin", SendPort::create(&mut ended.heap, stdin.channel)),
     ];
-    let environment = Struct::create_with_symbol_keys(&mut ended.heap, fields).into();
+    let environment = Struct::create_with_symbol_keys(&mut ended.heap, true, fields).into();
     let mut tracer = StackTracer::default();
-    let platform = HirId::create(&mut ended.heap, hir::Id::platform());
+    let platform = HirId::create(&mut ended.heap, true, hir::Id::platform());
     vm.initialize_for_function(
         ended.heap,
         ended.constant_mapping,