… compiler/vm/src/heap/mod.rs
… compiler/vm/src/heap/object.rs
… compiler/vm/src/heap/object_heap/hir_id.rs
… compiler/vm/src/heap/object_heap/int.rs
… compiler/vm/src/heap/object_heap/list.rs
… compiler/vm/src/heap/object_heap/mod.rs
… compiler/vm/src/heap/object_heap/struct_.rs
… compiler/vm/src/heap/object_heap/tag.rs
… compiler/vm/src/heap/object_heap/text.rs
… compiler/vm/src/heap/object_inline/int.rs
… compiler/vm/src/heap/object_representation.md
… compiler/vm/src/mir_to_lir.rs
… compiler/vm/src/vm.rs
… packages/examples/iterable.candy