<!-- Please enter the corresponding issue ID: -->


<!-- Please summarize your changes: -->
For a relatively short program (`packages/examples/iterable.canddy`, but with a longer text), the execution got slower:

```
Before: cargo run --release -- run packages/examples/iterable.candy
  Time (mean ± σ):     789.7 ms ±  42.1 ms    [User: 725.0 ms, System: 63.0 ms]
  Range (min … max):   717.9 ms … 849.0 ms    10 runs

After: cargo run --release -- run packages/examples/iterable.candy
  Time (mean ± σ):      1.028 s ±  0.082 s    [User: 0.968 s, System: 0.062 s]
  Range (min … max):    0.919 s …  1.208 s    10 runs
```

I'm guessing this is mostly because we now have to check whether an object is reference counted or not quite commonly. This effect should be reduced once we add a new IR for tracking dups and drops.

Edit: Our CI benchmarks improved: https://github.com/candy-lang/candy/commit/3021dcc619da6c85944311f687dcaf37091736e1#commitcomment-122395527


### Checklist
<!-- Please check if your PR fulfills the following requirements: -->

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)